Relocate GraphDatabasePlugin contract to Domain/GraphDatabase
